// server/enums/jobSector.ts
enum JobSector {
  OUTDOOR = 'OUTDOOR',
  CONSTRUCTION = 'CONSTRUCTION',
  DRIVING = 'DRIVING',
  BEAUTY = 'BEAUTY',
  HOSPITALITY = 'HOSPITALITY',
  TECHNICAL = 'TECHNICAL',
  MANUFACTURING = 'MANUFACTURING',
  OFFICE = 'OFFICE',
  RETAIL = 'RETAIL',
  SPORTS = 'SPORTS',
  WAREHOUSING = 'WAREHOUSING',
  EDUCATION_TRAINING = 'EDUCATION_TRAINING',
  WASTE_MANAGEMENT = 'WASTE_MANAGEMENT',
  CLEANING_AND_MAINTENANCE = 'CLEANING_AND_MAINTENANCE',
  OTHER = 'OTHER',
}

export default JobSector